refactor: implement do method for timer
This change adds a do method for timer (similar to ticker’s do method).
It also adds a test case to verifies that backwards compatibility is
preserved for a quirk in the current Observe and ticker implementation.

That is, a tick triggers the observeUnlocked method. I’m not aware of
any piece of code actually relying on this behavior, but it’s better to
change that separately if we want to.

Additionally, we introduce a small convention for Time’s methods: if the
method does not acquire the lock, it has Unlocked suffix. A notable
exception is the do method on timer and ticker that is expected to run
under Time’s lock when scheduled temporal effects are triggered.